:sparkles::feat:Validação de dados: usuário, professor, comentário e avaliação

+++ src/usuario/usuario.service.ts
@@ -1,11 +1,21 @@
-import { Injectable } from '@nestjs/common';
-import { Prisma, PrismaClient } from '../../generated/prisma';
+import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
+import { PrismaClient } from '../../generated/prisma';
+import { CreateUsuarioDto } from './dto/create-usuario.dto';
+import { UpdateUsuarioDto } from './dto/update-usuario.dto';
 
 @Injectable()
 export class UsuarioService {
   private prisma = new PrismaClient();
 
-  async create(data: Prisma.UsuarioCreateInput) {
+  async create(data: CreateUsuarioDto) {
+    const existente = await this.prisma.usuario.findUnique({
+      where: { email: data.email },
+    })
+
+  if (existente) {
+    throw new BadRequestException('Já existe um usuário com este email.');
+  }
+
     return this.prisma.usuario.create({ data });
   }
 
@@ -14,10 +24,34 @@ export class UsuarioService {
   }
 
   async findOne(id: number) {
-    return this.prisma.usuario.findUnique({ where: { id } });
+
+    //Verificando se o usuario existe
+    const usuario = await this.prisma.usuario.findUnique({ where: { id } });
+    if (!usuario) {
+      throw new NotFoundException('Usuário não encontrado.');
+    }
+
+    return usuario;
   }
 
-  async update(id: number, data: Prisma.UsuarioUpdateInput) {
+  async update(id: number, data: UpdateUsuarioDto) {
+
+    //Verificando se o usuario existe
+    const usuario = await this.prisma.usuario.findUnique({ where: { id } });
+    if (!usuario) {
+      throw new NotFoundException('Usuário não encontrado.');
+    }
+    
+    //Se estiver tentando alterar o email, verificar se ele já está em uso
+    if (data.email && data.email !== usuario.email) {
+      const emailExistente = await this.prisma.usuario.findUnique({
+        where: { email: data.email },
+      });
+      if (emailExistente) {
+        throw new BadRequestException('Este email já está em uso por outro usuário.');
+      }
+    }
+
     return this.prisma.usuario.update({
       where: { id },
       data,
@@ -25,6 +59,13 @@ export class UsuarioService {
   }
 
   async remove(id: number) {
+
+    //Verificando se o usuario existe
+    const usuario = await this.prisma.usuario.findUnique({ where: { id } });
+    if (!usuario) {
+      throw new NotFoundException('Usuário não encontrado.');
+    }
+    
     return this.prisma.usuario.delete({ where: { id } });
   }
 }
